docker-compose: Bump gunicorn timeout to 20 seconds
When a Zeek script is running longer than 10 seconds gunicorn times out
the POST /run request and the response the UI gets is HTTP 502 Bad Gateway.
There's no indication in the UI that this occurred and it looks like
execution is just hanging.

The worker terminates script execution properly after 15(+2) seconds
in runbro using the timeout command. Bump the gunicorn timeout to 20 seconds
to cover that plus a bit of wiggle room.