Add client-side syntax checking against Zeek tree-sitter grammar
This adds client-side validation of code via tree-sitter-zeek. In
contrast to errors from `zeek` itself errors here are marked warnings,
not errors. To support this we build the tree-sitter-zeek grammar into a
WASM file and load it with `web-tree-sitter`. To support `web-tree-sitter`
we need to customize the webpack setup which we can do by switching from
`react-scripts` to `craco`. The grammar itself is added as a Git
submodule.