Fix dependencies for zeek-format support
We previously used zeekscript-1.2.8 which did not pin its dependency on
tree-sitter which broke as that dependency moved along in an
incompatible way. This patch bumps us to a fixed version of zeekscript
(the latest). We also drop the dependency on tree-sitter which we never
used directly.

This fixes building the compose setup with `docker compose build` for me
(on macos with an aarch64 Docker).